test: ignore invalid test
This test is invalid, even though the filename suggests otherwise.
The expression: `var x = / [a - z /]$ / i;` is considered a
VariableStatement with two declarations, `x` and `i`.
It would be valid if `var x = / [a - z /]$ /i;` there was no space
between the last `/` and `i` - then it would be considered a modifier
and that case does work - as can be seen by the unit tests.